+++ api/forgot_test.go
@@ -22,6 +22,20 @@ func TestForgotResponds(t *testing.T) {
 			respCode: 200,
 		},
 		{
+			// always returns a 200 for patient
+			// with shortKey
+			method:   "POST",
+			url:      "/send/forgot/patient@myemail.com",
+			respCode: 200,
+		},
+		{
+			// always returns a 200 for patient
+			// without shortKey
+			method:   "POST",
+			url:      "/send/forgot/clinic@myemail.com",
+			respCode: 200,
+		},
+		{
 			// always returns a 200 if properly formed
 			returnNone: true,
 			method:     "POST",
@@ -35,6 +49,36 @@ func TestForgotResponds(t *testing.T) {
 			body: testJSONObject{
 				"key":      "1234_aK3yxxx123",
 				"email":    "me@myemail.com",
+				"password": "myN3wpa55w0rd",
+			},
+		},
+		{
+			method:   "PUT",
+			url:      "/accept/forgot",
+			respCode: 200,
+			body: testJSONObject{
+				"shortkey": "12345678",
+				"email":    "patient@myemail.com",
+				"password": "myN3wpa55w0rd",
+			},
+		},
+		{
+			method:   "PUT",
+			url:      "/accept/forgot",
+			respCode: 404,
+			body: testJSONObject{
+				"shortkey": "00000000",
+				"email":    "expired@myemail.com",
+				"password": "myN3wpa55w0rd",
+			},
+		},
+		{
+			method:   "PUT",
+			url:      "/accept/forgot",
+			respCode: 404,
+			body: testJSONObject{
+				"shortkey": "11111111",
+				"email":    "doesnotexist@myemail.com",
 				"password": "myN3wpa55w0rd",
 			},
 		},