+++ api/responsableMockGatekeeper.go
@@ -60,3 +60,7 @@ func (c *ResponsableMockGatekeeper) SetPermissions(userID, groupID string, permi
 	}
 	panic("SetPermissionsResponses unavailable")
 }
+
+func (c *ResponsableMockGatekeeper) GroupsForUser(userID string) (clients.UsersPermissions, error) {
+	panic("GroupsForUser unavailable")
+}