+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
+## Unreleased
+### Added
+- PT-1382: Reduce activation key length for password reset
+
 ## 0.9.0 - 2020-05-18
 ### Changed
 - PT-1278 Integrate Tidepool master for hydrophone